pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testrail-api-module"
version = "0.4.0"
description = "A comprehensive Python wrapper for the TestRail API with enhanced error handling and performance improvements"
authors = [
    { name = "Matt Troutman", email = "github@trtmn.com" },
    { name = "Christian Thompson", email = "example@example.com" }
]
requires-python = ">=3.11"
license = { file = "LICENSE" }
readme = "README.md"
dependencies = [
    "fastmcp>=0.9.0",
    "pytest>=8.4.2",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pdoc>=14.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "mypy>=1.0.0",
    "toml>=0.10.0",
    "python-dotenv>=1.0.0",
]

# Optional docs extra so consumers can install only what's needed to build docs
docs = [
    "pdoc>=14.0.0",
]

# Optional MCP extra for backward compatibility (MCP is now in base dependencies)
# This extra is kept for compatibility but is no longer needed
mcp = [
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Homepage = "https://github.com/trtmn/testrail_api_module"
Docs = 'https://trtmn.github.io/testrail_api_module/'

[project.scripts]
testrail-mcp-server = "testrail_api_module.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]



[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# Include stubs in the package
[tool.setuptools.package-data]
testrail_api_module = ["*.pyi", "py.typed"]

# Coverage configuration
[tool.coverage.run]
source = ["src/testrail_api_module"]
omit = [
    "*/utilities/*",
    "*/tests/*",
    "*/build/*",
    "*/dist/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
